+++ modules/impermanence.nix
@@ -1,0 +1,7 @@
+{
+  hmModule = { config, pkgs, lib, ... }: let
+    inherit (lib) mkIf;
+  in {
+    programs.vscode.mutableExtensionsDir = mkIf config.programs.vscode.enable false;
+  };
+}